Use new preproc and source position classes
